Tokens
Worked on outputting a token to use for messages.
Imported google-services.json to let us use Firebase to hopefully create tokens.
Created new class MyFirebaseMessagingService, edited build.gradle (root and app) and AndroidManifest.xml
Created MyFirebaseInstanceIDService but it is deprecated so it can probably be deleted soon.

Also finalized look of log in and sign up buttons so that they are long enough.