Committing for Zolboo
Since Zolboo is incapable of committing AND pulling down code I am doing this for him. He worked on notifications but they are still broken right now for some unknown reason, it appears as if it's not reading our code for when we receive a notification.

Time for some super sleuthing to fix this problem!!!! Go team