Notifications are back up
So, I'm not sure exactly where I'm committing this to. I just made a branch called dev but I don't think it's committing there.

Anyway, with this current code the notifications are working again. I re-added a getToken method, and I think that it was sending the notifications to the wrong token. (We need TWO getToken methods - one with a view parameter and one with no parameter.)

This is the code from 3/28. We're going to have to add back in everything that we've done since then. It shouldn't be that bad; it's only 2 or 3 days of code.

tl;dr:
NOTIFICATIONS ARE BACK!!!!