Notifications are working
Notifications are working again. I think it's connected to the getToken(View view) method. Without that, we aren't getting notifications. I think it also has to do with deleting the code calling getToken in activity_main.xml - it gives an error in getToken when it is deleted. So if there is a way to get rid of the getToken button and still get notifications, that would be great.

tl;dr:
NOTIFICATIONS ARE WORKING!