Committing for myself
Zolboo is now capable of committing AND pulling down code. WOW. Didnt expect that, did you?

Also the notifications are all fixed; we had to update Todd's side and
set the priority of the notification to HIGH and communicate through the Android Manifest.
Turns out we kinda didnt need the onMessageReceived method at all from
MyFirebaseMessagingService.java class. Poor class. Everybody thought he
was special and useful but he's really not. Just like everything else,
it is replaced, unable to withstand the test of time. Seems like
we are all some sort of modification of MyFirebaseMessagingService.java class
in some senses. Always wanting and always dreaming, yet never fulfilled.

Anyways, Olivia, sorry for not pulling down and committing last time. I
was laughing when I read that the first time because I thought you were
messing around but now that I read it again today, it seems like it was
more than just a joke. So I sincerely apologize for being a bad programming
partner.

What is the limit of this commit message?