Notification icon
WE HAVE A NOTIFICATION ICON!!!! I can't believe the day is finally here. Because our MyFirebaseMessagingService code isn't running for whatever reason, we have to set everything in the AndroidManifest file. I set it as the resource for Firebase's default notification icon and it is finally working.

Now notifications open up the app, have an icon, and have reworded text (including who the message was sent from). I think we're done with notifications!