Settings page
Added a link from MainActivity to a new settings page. It has links to edit profile, and log out, as well as links that take you to the about, privacy policy, and FAQ pages on the website.

Settings icon on MainActivity page takes you to settings page, and back arrow on Settings page takes you back to MainActivity.