Privacy policy page
Creating an inline privacy policy page so it doesn't redirect to the website. Privacy policy page should be public, but for some reason it was saying we had to log in to see it, so we're storing it in the app.

Next time I'll do the same thing for the FAQ page, just so it looks cleaner.